Move implementation of each crossover operator to separate module
