+++ src/ga.rs
@@ -320,15 +320,9 @@ where
 
             // FIXME: Do not assume that population size is an even number.
             self.timer.start();
-            for parents in mating_pool.chunks(2) {
-                let crt_children =
-                    self.config
-                        .crossover_operator
-                        .apply(&self.metadata, parents[0], parents[1]);
-
-                children.push(crt_children.0);
-                children.push(crt_children.1);
-            }
+            self.config
+                .crossover_operator
+                .apply(&self.metadata, &mating_pool, &mut children);
             self.metadata.crossover_dur = Some(self.timer.elapsed());
 
             self.timer.start();
