+++ src/ga/operators/crossover.rs
@@ -5,6 +5,7 @@ pub use impls::*;
 
 use crate::ga::individual::IndividualTrait;
 use crate::ga::GAMetadata;
+
 /// # Crossover Operator
 ///
 /// This trait defines common behaviour for crossover operators.
@@ -16,10 +17,30 @@ pub trait CrossoverOperator<IndividualT: IndividualTrait> {
     ///
     /// * `parent_1` - First parent to take part in recombination
     /// * `parent_2` - Second parent to take part in recombination
-    fn apply(
+    fn apply_legacy(
         &mut self,
         metadata: &GAMetadata,
         parent_1: &IndividualT,
         parent_2: &IndividualT,
     ) -> (IndividualT, IndividualT);
+
+    /// FIXME: Understand lifetimes here!
+    // fn apply_iter<'i, InputIter, OutputIter>(
+    //     &mut self,
+    //     metadata: &GAMetadata,
+    //     selected: InputIter,
+    // ) -> OutputIter
+    // where
+    //     InputIter: Iterator<Item = &'i IndividualT>,
+    //     OutputIter: Iterator<Item = IndividualT>,
+    //     IndividualT: 'i;
+
+    /// Apply crossover operator to the selected population part.
+    ///
+    /// ## Arguments
+    ///
+    /// * `metadata` - metadata provided by the GA runtime,
+    /// * `selected` - result of running selection operator,
+    /// * `output` - buffer for the individuals created by the crossover operator.
+    fn apply(&mut self, metadata: &GAMetadata, selected: &[&IndividualT], output: &mut Vec<IndividualT>);
 }
