+++ tests/crossover_tests.rs
@@ -14,7 +14,7 @@ fn operator_takes_values_from_parents<T: CrossoverOperator<RealValueIndividual>>
     let parents = RandomPoints::new(30).generate(2);
     assert_eq!(parents.len(), 2, "Expected population of size 2");
 
-    let (child_1, child_2) = operator.apply(&GAMetadata::default(), &parents[0], &parents[1]);
+    let (child_1, child_2) = operator.apply_legacy(&GAMetadata::default(), &parents[0], &parents[1]);
     for (i, (gene_1, gene_2)) in std::iter::zip(child_1.chromosome(), child_2.chromosome()).enumerate() {
         assert!(parents[0].chromosome()[i] == *gene_1 || parents[1].chromosome()[i] == *gene_1);
         assert!(parents[0].chromosome()[i] == *gene_2 || parents[1].chromosome()[i] == *gene_2);
@@ -50,7 +50,7 @@ fn ppx_test() {
 
     let p1 = Individual::from((0..10).collect_vec());
     let p2 = Individual::from((0..10).rev().collect_vec());
-    let (c1, c2) = op.apply(&GAMetadata::default(), &p1, &p2);
+    let (c1, c2) = op.apply_legacy(&GAMetadata::default(), &p1, &p2);
 
     c1.chromosome
         .iter()
